Moving stuff from SegmentPropagator and other utils class
