+++ drools-core/src/main/java/org/drools/core/common/TupleSets.java
@@ -18,6 +18,7 @@
  */
 package org.drools.core.common;
 
+import org.drools.core.reteoo.Tuple;
 import org.drools.core.reteoo.TupleImpl;
 
 public interface TupleSets {
@@ -55,4 +56,6 @@ public interface TupleSets {
 
     TupleImpl getNormalizedDeleteFirst();
     boolean addNormalizedDelete(TupleImpl leftTuple);
+    
+
 }
